api_db/requirements.txt
fastapi==0.95.2
fastavro==1.7.4
pandas==2.0.1
uvicorn==0.22.0
fsspec==2023.5.0
gcsfs==2023.5.0
google-cloud-storage==2.9.0
google-cloud-bigquery==3.10.0